Abstract target to allow for legacy & new formats
This is the first commit to abstract out the HtmlPublisherTarget and
duplicate the existing HtmlPublisherTarget for a HtmlPublisherNewTarget.

Doing this allows the existing legacy target to continue working as per
today whilst allowing changes to the new format that can then break with
the way that things are currently done.